Add searchBar and move renderedrooms to a separte component
