+++ src/components/Home/Home.js
@@ -1,11 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { Link } from 'react-router-dom';
 import { fetchRooms } from '../../redux/thunks/fetchRooms';
 import './home.css';
 import RoomFilters from './Filter';
-import CalculateRating from '../../utils/CalculateRating';
-import petroomImage from '../../images/petroom.jpg';
+import SearchBar from './SearchBar';
+import RenderedRooms from './RenderedRooms';
 
 function Home() {
   const dispatch = useDispatch();
@@ -13,6 +12,8 @@ function Home() {
 
   const [typeOfPetFilter, setTypeOfPetFilter] = useState('');
   const [sizeOfPetFilter, setSizeOfPetFilter] = useState('');
+
+  const [searchText, setSearchText] = useState('');
 
   useEffect(() => {
     dispatch(fetchRooms());
@@ -31,28 +32,16 @@ function Home() {
     );
   }
 
-  let renderedRooms = [];
+  const handleSearch = (event) => {
+    setSearchText(event.target.value);
+  };
+
   if (Array.isArray(rooms)) {
-    renderedRooms = filteredRooms.map((room) => (
-      <Link
-        key={room.id}
-        to={`pet_room/${room.id}`}
-        className="col-sm-5 mb-2 border no-style-link d-flex"
-      >
-        <div className="w-50">
-          <h2>{room.name}</h2>
-          <p>{`Type of pet living here: ${room.type_of_pet}`}</p>
-          <p>{`Max sized accepted: ${room.max_size_accepted}`}</p>
-          <p>{`User Owner: ${room.user_id}`}</p>
-          <p>{`Rating: ${CalculateRating(room.rating)}`}</p>
-          <p>{`Price: ${room.price}`}</p>
-        </div>
-        <div className="w-50">
-          <img src={petroomImage} alt="room-img" className="img-fluid" />
-        </div>
-      </Link>
-    ));
+    filteredRooms = filteredRooms.filter(
+      (room) => room.name.toLowerCase().includes(searchText.toLowerCase()),
+    );
   }
+
   return (
     <div className="container-fluid">
       <RoomFilters
@@ -61,7 +50,10 @@ function Home() {
         sizeOfPetFilter={sizeOfPetFilter}
         setSizeOfPetFilter={setSizeOfPetFilter}
       />
-      <div className="row justify-content-around">{renderedRooms}</div>
+      <SearchBar onChange={handleSearch} />
+      <div className="row justify-content-around">
+        <RenderedRooms rooms={filteredRooms} />
+      </div>
     </div>
   );
 }
